pyproject.toml
[project]
name = "investing-agents"
version = "0.1.0"
description = "A2A Server for Investment Strategy Agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.0",
    "httpx>=0.28.1",
    "pydantic>=2.11.0",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.2.2",
    "langchain-core>=0.3.0",
]

[project.optional-dependencies]
openai = ["langchain-openai>=0.3.0"]
azure = ["langchain-openai>=0.3.0"]
anthropic = ["langchain-anthropic>=0.3.0"]
google = ["langchain-google-genai>=2.0.0"]
ollama = ["langchain-ollama>=0.3.0"]
ui = ["streamlit>=1.44.0"]
all = [
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-ollama>=0.3.0",
]
all-ui = [
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-ollama>=0.3.0",
    "streamlit>=1.44.0",
]

[project.scripts]
investing-agents = "investing_agents.__main__:main"
investing-agents-ui = "investing_agents.streamlit_launcher:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/investing_agents"]

[dependency-groups]
dev = ["pytest>=9.0.3", "pytest-asyncio>=0.25.2", "ruff>=0.12.8"]
[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "dist",
    "build",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # Line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["investing_agents"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
